refactor: make scrape_review easy to use to test individual pages
The old method took the Beautiful Soup object of the review link
as an input and parsed the URL out. I refactored this to take the URL,
so now the scrape page method parses the URL for each review and passes
that as the input.

This is useful because now I can debug the remaining parsing errors
that I see by running them on that single URL instead of having to
run a big list of pages again.