bug: fix multiprocessing review format errors
I was using class attributes to save the format of a review, which
was being changed by the various workers and leading to parsing errors.

The method that determines the review format now returns a dictionary
with the info instead of saving it on the class. This way many processes
can run at one time.