bug: ensure each data file is unique and no data is missed
Since the data attribute is shared on the instance, when saving
multiple files there were race conditions popping up and some data
was being missed.

Each page now has its own list of review data, which is saved when
the page is done. This will lead to a large number of files, but
it will keep the scrape time speedy. I can go back after the scrape is
done and easily combine the files.